refactor(game): improve structure with non-blocking state handling

+++ src/hardware/Buttons/ButtonManager.h
@@ -8,7 +8,7 @@ class ButtonManager {
 
         bool lastReading[PINS_COUNT] {};
         bool stableState[PINS_COUNT] {};
-        unsigned long long lastChangeTime[PINS_COUNT] {};
+        unsigned long lastChangeTime[PINS_COUNT] {};
 
         bool detectPress(int index);
